backend/requirements.txt
fastapi
uvicorn
numpy
scikit-learn
rank_bm25
sentence-transformers
faiss-cpu
torch